fix test
The test originally assumed, just like the standalone application,
that all available routes that don't contain the word "sso" are
actually applications secured using the SSO server, and hence subject
to testing.

That doesn't necessarily have to be true. E.g. in our CI environment,
there's a Jenkins server running, and the CI dashboard application.
This commit improves how applications to test are selected.
Instead of taking all routes that don't contain the word "sso",
it takes all routes containing the word "secured". Of course that's
not ideal, and we should probably move to selecting by labels,
but it's much better than the original way.